+++ projects/OG-Financial/src/main/java/com/opengamma/financial/analytics/volatility/surface/BloombergEquityFutureOptionVolatilitySurfaceInstrumentProvider.java
@@ -15,7 +15,7 @@ import org.threeten.bp.format.DateTimeFormatter;
 
 import com.opengamma.financial.analytics.ircurve.NextExpiryAdjuster;
 import com.opengamma.financial.analytics.model.FutureOptionExpiries;
-import com.opengamma.financial.convention.ExchangeTradedInstrumentExpiryCalculator;
+import com.opengamma.financial.convention.expirycalc.ExchangeTradedInstrumentExpiryCalculator;
 import com.opengamma.id.ExternalId;
 import com.opengamma.util.ArgumentChecker;