+++ projects/OG-Financial/src/main/java/com/opengamma/financial/convention/expirycalc/BondFutureExpiryCalculator.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.financial.convention;
+package com.opengamma.financial.convention.expirycalc;
 
 import org.threeten.bp.DayOfWeek;
 import org.threeten.bp.LocalDate;
@@ -14,14 +14,14 @@ import com.opengamma.financial.convention.calendar.Calendar;
 import com.opengamma.util.ArgumentChecker;
 
 /**
- * Expiry calculator for bond future options.
+ * Expiry calculator for bond futures.
  */
-public final class BondFutureOptionExpiryCalculator implements ExchangeTradedInstrumentExpiryCalculator {
+public final class BondFutureExpiryCalculator implements ExchangeTradedInstrumentExpiryCalculator {
 
   /** Name of the calculator */
   public static final String NAME = "BondFutureOptionExpiryCalculator";
   /** Singleton. */
-  private static final BondFutureOptionExpiryCalculator INSTANCE = new BondFutureOptionExpiryCalculator();
+  private static final BondFutureExpiryCalculator INSTANCE = new BondFutureExpiryCalculator();
   /** Adjuster. */
   private static final TemporalAdjuster LAST_DAY_ADJUSTER = TemporalAdjusters.lastDayOfMonth();
   /** Adjuster. */
@@ -34,14 +34,14 @@ public final class BondFutureOptionExpiryCalculator implements ExchangeTradedIns
    * 
    * @return the instance, not null
    */
-  public static BondFutureOptionExpiryCalculator getInstance() {
+  public static BondFutureExpiryCalculator getInstance() {
     return INSTANCE;
   }
 
   /**
    * Restricted constructor.
    */
-  private BondFutureOptionExpiryCalculator() {
+  private BondFutureExpiryCalculator() {
   }
 
   //-------------------------------------------------------------------------
@@ -54,10 +54,10 @@ public final class BondFutureOptionExpiryCalculator implements ExchangeTradedIns
     LocalDate lastDayOfMonth;
     LocalDate lastFridayOfMonth;
     if (today.isAfter(lastFridayOfThisMonth)) {
-      lastDayOfMonth = today.plusMonths(n).with(LAST_DAY_ADJUSTER);
+      lastDayOfMonth = today.plusMonths(3 * n).with(LAST_DAY_ADJUSTER);
       lastFridayOfMonth = lastDayOfMonth.with(PREVIOUS_OR_CURRENT_FRIDAY_ADJUSTER);
     } else {
-      lastDayOfMonth = today.plusMonths(n - 1).with(LAST_DAY_ADJUSTER);
+      lastDayOfMonth = today.plusMonths(3 * (n - 1)).with(LAST_DAY_ADJUSTER);
       lastFridayOfMonth = lastDayOfMonth.with(PREVIOUS_OR_CURRENT_FRIDAY_ADJUSTER);
     }
     int nBusinessDays = 0;