+++ projects/OG-Financial/src/main/java/com/opengamma/financial/convention/expirycalc/LiveCattleFutureOptionExpiryCalculator.java
@@ -3,10 +3,11 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.financial.convention;
+package com.opengamma.financial.convention.expirycalc;
 
 import java.util.Arrays;
 
+import org.threeten.bp.DayOfWeek;
 import org.threeten.bp.LocalDate;
 import org.threeten.bp.Month;
 import org.threeten.bp.temporal.TemporalAdjuster;
@@ -16,18 +17,18 @@ import com.opengamma.financial.convention.calendar.Calendar;
 import com.opengamma.util.ArgumentChecker;
 
 /**
- * Expiry calculator for live cattle futures.
+ * Expiry calculator for live cattle future options.
  */
-public final class LiveCattleFutureExpiryCalculator implements ExchangeTradedInstrumentExpiryCalculator {
+public final class LiveCattleFutureOptionExpiryCalculator implements ExchangeTradedInstrumentExpiryCalculator {
 
   /** Name of the calculator */
-  public static final String NAME = "LiveCattleFutureExpiryCalculator";
+  public static final String NAME = "LiveCattleFutureOptionExpiryCalculator";
   /** Singleton. */
-  private static final LiveCattleFutureExpiryCalculator INSTANCE = new LiveCattleFutureExpiryCalculator();
+  private static final LiveCattleFutureOptionExpiryCalculator INSTANCE = new LiveCattleFutureOptionExpiryCalculator();
   /** Adjuster. */
-  private static final TemporalAdjuster LAST_DAY_ADJUSTER = TemporalAdjusters.lastDayOfMonth();
+  private static final TemporalAdjuster FIRST_FRIDAY_ADJUSTER = TemporalAdjusters.firstInMonth(DayOfWeek.FRIDAY);
   /** Months when futures expire. */
-  private static final Month[] CATTLE_FUTURE_EXPIRY_MONTHS = {
+  private static final Month[] CATTLE_OPTION_EXPIRY_MONTHS = {
     Month.FEBRUARY, Month.APRIL, Month.JUNE, Month.AUGUST, Month.OCTOBER, Month.DECEMBER
   };
 
@@ -36,21 +37,21 @@ public final class LiveCattleFutureExpiryCalculator implements ExchangeTradedIns
    * 
    * @return the instance, not null
    */
-  public static LiveCattleFutureExpiryCalculator getInstance() {
+  public static LiveCattleFutureOptionExpiryCalculator getInstance() {
     return INSTANCE;
   }
 
   /**
    * Restricted constructor.
    */
-  private LiveCattleFutureExpiryCalculator() {
+  private LiveCattleFutureOptionExpiryCalculator() {
   }
 
   //-------------------------------------------------------------------------
   /**
-   * Expiry date of Soybean Futures:
-   * The last business day of the contract month.
-   * See http://www.cmegroup.com/trading/agricultural/livestock/live-cattle_contractSpecs_futures.html
+   * Expiry date of Soybean Future Options:
+   * The first Friday of the month.
+   * See http://www.cmegroup.com/trading/agricultural/livestock/live-cattle_contractSpecs_options.html
    * 
    * @param n  the n'th expiry date after today, greater than zero
    * @param today  the valuation date, not null
@@ -63,34 +64,40 @@ public final class LiveCattleFutureExpiryCalculator implements ExchangeTradedIns
     ArgumentChecker.notNull(today, "today");
     ArgumentChecker.notNull(holidayCalendar, "holiday calendar");
 
-    LocalDate expiryDate = getExpiryMonth(n, today);
-
-    while (!holidayCalendar.isWorkingDay(expiryDate)) {
-      expiryDate = expiryDate.minusDays(1);
+    LocalDate expiry = getExpiryMonth(n, today).with(FIRST_FRIDAY_ADJUSTER);
+    while (!holidayCalendar.isWorkingDay(expiry)) {
+      expiry = expiry.minusDays(1);
     }
-    return expiryDate;
+    return expiry;
   }
 
   @Override
   public LocalDate getExpiryMonth(final int n, final LocalDate today) {
     ArgumentChecker.isTrue(n > 0, "n must be greater than zero");
     ArgumentChecker.notNull(today, "today");
-    LocalDate expiryDate = today;
-    for (int m = n; m > 0; m--) {
-      expiryDate = getNextExpiryMonth(expiryDate);
+    // There are 1 serial options
+    if (n < 2) {
+      return today.plusMonths(n);
     }
-    return expiryDate.with(LAST_DAY_ADJUSTER);
+    // now
+    int m = n - 1;
+    LocalDate expiryDate = today.plusMonths(1);
+    while (m > 0) {
+      expiryDate = getNextExpiryMonth(expiryDate);
+      m--;
+    }
+    return expiryDate;
   }
 
   private LocalDate getNextExpiryMonth(final LocalDate dtCurrent) {
     Month mthCurrent = dtCurrent.getMonth();
-    int idx = Arrays.binarySearch(CATTLE_FUTURE_EXPIRY_MONTHS, mthCurrent);
-    if (Math.abs(idx) >= (CATTLE_FUTURE_EXPIRY_MONTHS.length - 1)) {
-      return LocalDate.of(dtCurrent.getYear() + 1, Month.FEBRUARY, dtCurrent.getDayOfMonth());
+    int idx = Arrays.binarySearch(CATTLE_OPTION_EXPIRY_MONTHS, mthCurrent);
+    if (idx >= (CATTLE_OPTION_EXPIRY_MONTHS.length - 1)) {
+      return LocalDate.of(dtCurrent.getYear() + 1, CATTLE_OPTION_EXPIRY_MONTHS[0], dtCurrent.getDayOfMonth());
     } else if (idx >= 0) {
-      return dtCurrent.with(CATTLE_FUTURE_EXPIRY_MONTHS[idx + 1]);
+      return dtCurrent.with(CATTLE_OPTION_EXPIRY_MONTHS[idx + 1]);
     } else {
-      return dtCurrent.with(CATTLE_FUTURE_EXPIRY_MONTHS[-idx - 1]); //check this
+      return dtCurrent.with(CATTLE_OPTION_EXPIRY_MONTHS[-1 - idx]);
     }
   }